requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
albumentations==1.4.1
attrs==23.2.0
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.2.0
idna==3.6
imageio==2.34.0
Jinja2==3.1.3
joblib==1.3.2
lazy_loader==0.3
lightning==2.2.1
lightning-utilities==0.10.1
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numpy==1.26.4
opencv-python
opencv-python-headless
packaging==24.0
pandas==2.2.1
pillow
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.1
pytz==2024.1
PyYAML
scikit-image==0.22.0
scipy==1.12.0
six==1.16.0
sympy==1.12
threadpoolctl==3.3.0
tifffile==2024.2.12
torch
torchmetrics
torchvision
tqdm
triton==2.2.0
typing_extensions==4.10.0
tzdata==2024.1
yarl==1.9.4
segmentation_models_pytorch